+++ helper/ImageHelper.php
@@ -52,22 +52,28 @@ class ImageHelper {
 	 * @return string The cached Image URL
 	 */
 	public static function getLocalCacheImageUriForRemoteImage($cacheType = null, $remoteImageUrl = null) {
-		$explodedImageUrl = \explode('/', $remoteImageUrl);
-		$imageFilename = \end($explodedImageUrl);
-		$returnValue = CacheHelper::getImageCacheUri() . $cacheType . '/' . $imageFilename;
+		$themeOptions = \get_option('eve_theme_options', ThemeHelper::getThemeDefaultOptions());
+		$returnValue = $remoteImageUrl;
 
-		// if we don't have the image cached already
-		if(CacheHelper::checkCachedImage($cacheType, $imageFilename) === false) {
-			/**
-			 * Check if the content dir is writable and cache the image.
-			 * Otherwise set the remote image as return value.
-			 */
-			if(\is_dir(CacheHelper::getImageCacheDir() . $cacheType) && \is_writable(CacheHelper::getImageCacheDir() . $cacheType)) {
-				CacheHelper::cacheRemoteImageFile($cacheType, $remoteImageUrl);
-			} else {
-				$returnValue = $remoteImageUrl;
-			} // END if(\is_writable(\WP_CONTENT_DIR))
-		} // END if(CacheHelper::checkCachedImage($cacheType, $imageName) === false)
+		// Check if we should use image cache
+		if(!empty($themeOptions['cache']['remote-image-cache'])) {
+			$explodedImageUrl = \explode('/', $remoteImageUrl);
+			$imageFilename = \end($explodedImageUrl);
+			$returnValue = CacheHelper::getImageCacheUri() . $cacheType . '/' . $imageFilename;
+
+			// if we don't have the image cached already
+			if(CacheHelper::checkCachedImage($cacheType, $imageFilename) === false) {
+				/**
+				 * Check if the content dir is writable and cache the image.
+				 * Otherwise set the remote image as return value.
+				 */
+				if(\is_dir(CacheHelper::getImageCacheDir() . $cacheType) && \is_writable(CacheHelper::getImageCacheDir() . $cacheType)) {
+					CacheHelper::cacheRemoteImageFile($cacheType, $remoteImageUrl);
+				} else {
+					$returnValue = $remoteImageUrl;
+				} // END if(\is_dir(CacheHelper::getImageCacheDir() . $cacheType) && \is_writable(CacheHelper::getImageCacheDir() . $cacheType))
+			} // END if(CacheHelper::checkCachedImage($cacheType, $imageName) === false)
+		} // END if(!empty($themeOptions['cache']['remote-image-cache']))
 
 		return $returnValue;
 	} // END public static function getLocalCacheImageUri($cacheType = null, $remoteImageUrl = null)